@@ -1,75 +1,64 @@
 package org.springframework.samples.petclinic.api.application;
 
+import okhttp3.mockwebserver.MockResponse;
+import okhttp3.mockwebserver.MockWebServer;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.cloud.client.circuitbreaker.EnableCircuitBreaker;
-import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.http.MediaType;
-import org.springframework.samples.petclinic.api.dto.VisitDetails;
-import org.springframework.test.context.junit.jupiter.SpringJUnitConfig;
-import org.springframework.test.web.client.MockRestServiceServer;
-import org.springframework.web.client.RestTemplate;
+import org.springframework.samples.petclinic.api.dto.Visits;
+import org.springframework.web.reactive.function.client.WebClient;
+import reactor.core.publisher.Mono;
 
+import java.io.IOException;
 import java.util.Collections;
-import java.util.List;
-import java.util.Map;
+import java.util.function.Consumer;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.springframework.test.web.client.match.MockRestRequestMatchers.requestTo;
-import static org.springframework.test.web.client.response.MockRestResponseCreators.withServerError;
-import static org.springframework.test.web.client.response.MockRestResponseCreators.withSuccess;
 
-@EnableCircuitBreaker
-@EnableAspectJAutoProxy
-@SpringJUnitConfig(classes = {VisitsServiceClient.class, RestTemplate.class})
-class VisitsServiceClientTest {
+class VisitsServiceClientIntegrationTest {
 
     private static final Integer PET_ID = 1;
 
-    @Autowired
     private VisitsServiceClient visitsServiceClient;
 
-    @Autowired
-    private RestTemplate restTemplate;
-
-    private MockRestServiceServer mockServer;
+    private MockWebServer server;
 
     @BeforeEach
-    public void setUp() {
-        mockServer = MockRestServiceServer.createServer(restTemplate);
+    void setUp() {
+        server = new MockWebServer();
+        visitsServiceClient = new VisitsServiceClient(WebClient.builder());
+        visitsServiceClient.setHostname(server.url("/").toString());
+    }
+
+    @AfterEach
+    void shutdown() throws IOException {
+        this.server.shutdown();
     }
 
     @Test
-    public void getVisitsForPets_withAvailableVisitsService() {
-        mockServer.expect(requestTo("http://visits-service/pets/visits?petId=1"))
-            .andRespond(withSuccess("{\"items\":[{\"id\":5,\"date\":\"2018-11-15\",\"description\":\"test visit\",\"petId\":1}]}", MediaType.APPLICATION_JSON));
+    void getVisitsForPets_withAvailableVisitsService() {
+        prepareResponse(response -> response
+            .setHeader("Content-Type", "application/json")
+            .setBody("{\"items\":[{\"id\":5,\"date\":\"2018-11-15\",\"description\":\"test visit\",\"petId\":1}]}"));
 
-        Map<Integer, List<VisitDetails>> visits = visitsServiceClient.getVisitsForPets(Collections.singletonList(1));
+        Mono<Visits> visits = visitsServiceClient.getVisitsForPets(Collections.singletonList(1));
 
-        assertVisitDescriptionEquals(visits, PET_ID,"test visit");
+        assertVisitDescriptionEquals(visits.block(), PET_ID,"test visit");
     }
 
-    /**
-     * Test Hystrix fallback method
-     */
-    @Test
-    public void getVisitsForPets_withServerError() {
 
-        mockServer.expect(requestTo("http://visits-service/pets/visits?petId=1"))
-            .andRespond(withServerError());
-
-        Map<Integer, List<VisitDetails>> visits = visitsServiceClient.getVisitsForPets(Collections.singletonList(1));
-
-        assertEquals(0, visits.size());
+    private void assertVisitDescriptionEquals(Visits visits, int petId, String description) {
+        assertEquals(1, visits.getItems().size());
+        assertNotNull(visits.getItems().get(0));
+        assertEquals(petId, visits.getItems().get(0).getPetId());
+        assertEquals(description, visits.getItems().get(0).getDescription());
     }
 
-    private void assertVisitDescriptionEquals(Map<Integer, List<VisitDetails>> visits, int petId, String description) {
-        assertEquals(1, visits.size());
-        assertNotNull(visits.get(1));
-        assertEquals(1, visits.get(petId).size());
-        assertEquals(description, visits.get(petId).get(0).getDescription());
+    private void prepareResponse(Consumer<MockResponse> consumer) {
+        MockResponse response = new MockResponse();
+        consumer.accept(response);
+        this.server.enqueue(response);
     }
 
 }
